Change compaction priority from long to short (#2094)

+++ server/tserver/src/main/java/org/apache/accumulo/tserver/compactions/ExternalCompactionExecutor.java
@@ -190,10 +190,10 @@ public class ExternalCompactionExecutor implements CompactionExecutor {
   }
 
   public Stream<TCompactionQueueSummary> summarize() {
-    HashSet<Long> uniqPrios = new HashSet<Long>();
+    HashSet<Short> uniqPrios = new HashSet<Short>();
     queuedTask.forEach(task -> uniqPrios.add(task.getJob().getPriority()));
 
-    Stream<Long> prioStream = uniqPrios.stream();
+    Stream<Short> prioStream = uniqPrios.stream();
 
     if (uniqPrios.size() > 100) {
       // Until #2094 is addressed limit what is sent to the coordinator to avoid causing it run out
